# contracts/http/swagger.yaml
basePath: /
definitions:
...
  model.ActionExecutionStatus:
    enum:
    - pending
    - running
    - succeeded
    - failed
    - retrying
    - skipped
    - cancelled
    type: string
    x-enum-varnames:
    - ActionExecutionPending
    - ActionExecutionRunning
    - ActionExecutionSucceeded
    - ActionExecutionFailed
    - ActionExecutionRetrying
    - ActionExecutionSkipped
    - ActionExecutionCancelled
  model.ActionType:
    enum:
    - send_dm
    - timeout_user
    - kick_user
    - ban_user
    - remove_timeout
    - unban_user
    type: string
    x-enum-varnames:
    - ActionSendDM
    - ActionTimeoutUser
    - ActionKickUser
    - ActionBanUser
    - ActionRemoveTimeout
    - ActionUnbanUser
...
  routes.failedActionEnvelope:
    properties:
      action:
        $ref: '#/definitions/routes.failedActionResponse'
    type: object
  routes.failedActionListResponse:
    properties:
      executions:
        items:
          $ref: '#/definitions/routes.failedActionResponse'
        type: array
      total:
        type: integer
    type: object
  routes.failedActionResponse:
    properties:
      action_type:
        $ref: '#/definitions/model.ActionType'
      attempt_count:
        type: integer
      case_id:
        type: string
      created_at:
        type: string
      id:
        type: string
      last_error:
        type: string
      last_error_code:
        type: string
      max_retries:
        type: integer
      safe_for_retry:
        type: boolean
      status:
        $ref: '#/definitions/model.ActionExecutionStatus'
      updated_at:
        type: string
    type: object
...
  routes.reverseActionRequest:
    properties:
      action_type:
        $ref: '#/definitions/model.ActionType'
      appeal_id:
        type: string
      confirm:
        type: boolean
      original_execution_id:
        type: string
    type: object
...
info:
  contact: {}
  description: HTTP boundary exposed by the Quack v5 backend.
  title: Quack HTTP API
  version: "5.0"
paths:
...
  /guilds/{discordGuildID}/action-failures:
    get:
      parameters:
      - description: Discord guild ID
        in: path
        name: discordGuildID
        required: true
        type: string
      - description: Page size
        in: query
        name: limit
        type: integer
      - description: Page offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/routes.failedActionListResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
      security:
      - CookieAuth: []
      summary: List failed case actions
      tags:
      - Action recovery
  /guilds/{discordGuildID}/action-failures/{executionID}/dismiss:
    post:
      parameters:
      - description: Discord guild ID
        in: path
        name: discordGuildID
        required: true
        type: string
      - description: Action execution ID
        in: path
        name: executionID
        required: true
        type: string
      - description: Retry-safe request key
        in: header
        name: Idempotency-Key
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/routes.failedActionEnvelope'
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - CookieAuth: []
      summary: Dismiss a failed case action
      tags:
      - Action recovery
  /guilds/{discordGuildID}/action-failures/{executionID}/retry:
    post:
      parameters:
      - description: Discord guild ID
        in: path
        name: discordGuildID
        required: true
        type: string
      - description: Action execution ID
        in: path
        name: executionID
        required: true
        type: string
      - description: Retry-safe request key
        in: header
        name: Idempotency-Key
        required: true
        type: string
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/routes.failedActionEnvelope'
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - CookieAuth: []
      summary: Retry a failed case action
      tags:
      - Action recovery
...
  /guilds/{discordGuildID}/cases/{caseRef}/reversals:
    post:
      consumes:
      - application/json
      parameters:
      - description: Discord guild ID
        in: path
        name: discordGuildID
        required: true
        type: string
      - description: Case ID or number
        in: path
        name: caseRef
        required: true
        type: string
      - description: Retry-safe request key
        in: header
        name: Idempotency-Key
        required: true
        type: string
      - description: Confirmed reversal
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/routes.reverseActionRequest'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - CookieAuth: []
      summary: Reverse a case action
      tags:
      - Action recovery
...
schemes:
- http
- https
securityDefinitions:
  CookieAuth:
    in: header
    name: Cookie
    type: apiKey
  MetricsKey:
    in: header
    name: X-Quack-Metrics-Key
    type: apiKey
  OpsKey:
    in: header
    name: X-Quack-Ops-Key
    type: apiKey
swagger: "2.0"